Added Weapon Equiping Functionality

+++ Assets/Resources/Scripts/Interfaces/ICanDamage.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface ICanDamage
 {
-    int Damage();
+    float Damage();
 }
